# app/static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Mechanic Shop",
    "version": "1.0.0",
    "description": "API for managing mechanics, customers, service tickets and inventory."
  },
  "basePath": "/",
  "schemes": ["http"],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT Authorization header using the Bearer scheme. Example: 'Authorization: Bearer {token}'"
    }
  },
...
  "paths": {
...
    "/inventory/": {
      "post": {
        "tags": ["Inventory"],
        "summary": "Add New Inventory Part",
        "description": "Adds a new part to the inventory.",
        "security": [{ "Bearer": [] }],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/InventoryPartPayload" }
          }
        ],
        "responses": {
          "201": { "description": "Part added to inventory", "schema": { "$ref": "#/definitions/InventoryPartResponse" } },
          "400": { "description": "Invalid request" }
        }
      },
      "get": {
        "tags": ["Inventory"],
        "summary": "Get All Inventory Parts",
        "description": "Anyone can see all parts in the inventory with pagination.",
        "parameters": [
          { "name": "page", "in": "query", "type": "integer", "required": false },
          { "name": "per_page", "in": "query", "type": "integer", "required": false }
        ],
        "responses": {
          "200": {
            "description": "List of inventory parts",
            "schema": { "type": "array", "items": { "$ref": "#/definitions/PublicPartResponse" } }
          }
        }
      }
    },
    "/inventory/low-stock": {
      "get": {
        "tags": ["Inventory"],
        "summary": "Get Low Stock Inventory Parts (Mechanic)",
        "description": "Mechanic can see all parts in the inventory that are low in stock. (threshold less than 10)",
        "security": [{ "Bearer": [] }],
        "parameters": [
          { 
            "name": "threshold", 
            "in": "query", 
            "type": "integer", 
            "required": false, 
            "description": "Show parts with stock quantity less than or equal to this number (10)." 
          }
        ],
        "responses": {
          "200": {
            "description": "List of low stock inventory parts",
            "schema": { 
              "type": "array", 
              "items": { "$ref": "#/definitions/InventoryPartResponse" } 
            }
          }
        }
      }
    },
    "/inventory/mechanic": {
      "get": {
        "tags": ["Inventory"],
        "summary": "Get All Inventory Parts (Mechanic)",
        "description": "Mechanic can see all parts in the inventory with pagination.",
        "security": [{ "Bearer": [] }],
        "parameters": [
          { "name": "page", "in": "query", "type": "integer", "required": false },
          { "name": "per_page", "in": "query", "type": "integer", "required": false }
        ],
        "responses": {
          "200": {
            "description": "List of inventory parts",
            "schema": { "type": "array", "items": { "$ref": "#/definitions/InventoryPartResponse" } }
          }
        }
      }
    },
    "/inventory/mechanic/search": {
      "get": {
        "tags": ["Inventory"],
        "summary": "Get Inventory Part by name (Mechanic)",
        "description": "Mechanic can search for a specific part in the inventory by name.",
        "security": [{ "Bearer": [] }],
        "parameters": [
          { 
            "name": "qPartName", 
            "in": "query", 
            "required": true, 
            "type": "string",
            "description": "Inventory Part name"
            }
          ],
        "responses": {
          "200": {
            "description": "Inventory Part name found",
            "schema": {
              "type": "array",
              "items": { "$ref": "#/definitions/PublicPartResponse" }
            }
          },
          "404": { "description": "Part not found" }
        }
      }
    },
    "/inventory/mechanic/{id}": {
      "get": {
        "tags": ["Inventory"],
        "summary": "Get an Inventory Part by ID (Mechanic)",
        "description": "Mechanic can search for a specific part in the inventory by ID.",
        "security": [{ "Bearer": [] }],
        "parameters": [
          { 
            "name": "id", 
            "in": "path",
            "required": true, 
            "type": "integer",
            "description": "Inventory Part ID"
            }
          ],
        "responses": {
          "200": {
            "description": "Inventory Part found",
            "schema": { "$ref": "#/definitions/InventoryPartResponse" }
          },
          "404": { "description": "Part not found" }
        }
      }
    },
    "/inventory/search": {
      "get": {
        "tags": ["Inventory"],
        "summary": "Get Inventory Part by name (Public)",
        "description": "Anyone can search for a specific part in the inventory by name.",
        "parameters": [
          { 
            "name": "qPartName", 
            "in": "query", 
            "required": true, 
            "type": "string",
            "description": "Inventory Part name"
            }
          ],
        "responses": {
          "200": {
            "description": "Inventory Part name found",
            "schema": {
              "type": "array",
              "items": { "$ref": "#/definitions/PublicPartResponse" }
            }
          },
          "404": { "description": "Part not found" }
        }
      }
    },
    "/inventory/{id}": {
      "get": {
        "tags": ["Inventory"],
        "summary": "Get Inventory Part by ID (Public)",
        "description": "Anyone can search for a specific part in the inventory by ID.",
        "parameters": [
          { 
            "name": "id", 
            "in": "path", 
            "required": true, 
            "type": "integer",
            "description": "Inventory Part ID"
            }
          ],
        "responses": {
          "200": {
            "description": "Inventory Part found",
            "schema": { "$ref": "#/definitions/PublicPartResponse" }
          },
          "404": { "description": "Part not found" }
        }
      },
      "put": {
        "tags": ["Inventory"],
        "summary": "Update an Inventory Part by ID",
        "description": "Mechanic updates an inventory part by ID.",
        "security": [{ "Bearer": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "type": "integer" },
          { "in": "body", "name": "body", "schema": { "$ref": "#/definitions/InventoryPartPayload" } }
        ],
        "responses": {
          "200": { "description": "Inventory part updated", "schema": { "$ref": "#/definitions/InventoryPartResponse" } },
          "400": { "description": "Invalid input" },
          "404": { "description": "Part not found" }
        }
      },
      "delete": {
        "tags": ["Inventory"],
        "summary": "Delete an Inventory Part by ID",
        "description": "Mechanic deletes an inventory part by ID.",
        "security": [{ "Bearer": [] }],
        "parameters": [
          { 
            "name": "id", 
            "in": "path", 
            "required": true, 
            "type": "integer", 
            "description": "Inventory Part ID"
          }
        ],
        "responses": {
          "204": { "description": "Inventory part deleted" },
          "404": { "description": "Part not found" }
        }
      }
    }
  },
  "definitions": {
...
    "InventoryPartResponse": {
      "type": "object",
      "properties": {
        "id": { "type": "integer" },
        "name": { "type": "string" },
        "quantity": { "type": "integer" },
        "price": { "type": "number", "format": "float" }
      }
    },
    "InventoryPartPayload": {
      "type": "object",
      "properties": {
        "name": { "type": "string" },
        "quantity": { "type": "integer" },
        "price": { "type": "number", "format": "float" }
      },
  "required": ["name", "quantity", "price"]
    },
    "PublicPartResponse": {
      "type": "object",
      "properties": {
        "id": { "type": "integer" },
        "name": { "type": "string" },
        "quantity": { "type": "integer" }
      }
    }
  }
}